move spend_sim into tests
