+++ chia/_tests/fee_estimation/test_mempoolitem_height_added.py
@@ -6,7 +6,7 @@ from typing import Callable, Optional
 import pytest
 from chia_rs import Coin, G2Element
 
-from chia.clvm.spend_sim import SimClient, SpendSim, sim_and_client
+from chia._tests.util.spend_sim import SimClient, SpendSim, sim_and_client
 from chia.consensus.constants import ConsensusConstants
 from chia.consensus.default_constants import DEFAULT_CONSTANTS
 from chia.full_node.bitcoin_fee_estimator import BitcoinFeeEstimator
